event_management/requirements.txt
asgiref==3.6.0
Django==4.2.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
python-decouple
psycopg2
qrcode
pillow
djangorestframework_simplejwt
PyJWT
cloudinary
django-cloudinary-storage
django-jazzmin
drf-yasg
django-debug-toolbar
# django_cloudinary_storage